+++ gd-smartfarm/src/index.jsx
@@ -1,7 +1,28 @@
 // @flow
-import reducers from 'gd-core/src/reducers';
 import render from 'gd-core/src/render';
 
+// $FlowFixMe
+import { addEndpoints, fetchSensors as __old_fetchSensors } from 'gd-core__old/app/actions';
+import { updateGeoStreamingConfig } from 'gd-geostreaming/src/actions/config';
+import { fetchParameters } from 'gd-geostreaming/src/actions/parameters';
+import { fetchSensors } from 'gd-geostreaming/src/actions/sensors';
 import routes from './routes';
+import reducers from './reducers';
 
-render(reducers, routes);
+import config from './config';
+
+const old_config = require('./old_config');
+
+window.configruntime = old_config;
+
+render(
+    reducers,
+    routes,
+    (store) => {
+        store.dispatch(updateGeoStreamingConfig(config));
+        store.dispatch(addEndpoints());
+        store.dispatch(__old_fetchSensors(config.geostreamingEndpoint));
+        store.dispatch(fetchParameters());
+        store.dispatch(fetchSensors());
+    }
+);