+++ gd-smartfarm/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Added header-footer layout and empty components for each page. [SMARTFARM-8](https://opensource.ncsa.illinois.edu/jira/browse/SMARTFARM-8)
+- Setup Geostreaming Pages and redux store (Explore and Download).  [SMARTFARM-12](https://opensource.ncsa.illinois.edu/jira/browse/SMARTFARM-12)
 
 ## [0.1.0] - 2020-10-01
 ### Added